+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add the Main contact task to the Transfer project task list
 - Show the main contact for a project in the project CSV exports
 - The all conditions met task for a transfer has updated content.
+- Add deeds of termination for MFA task to Transfer projects
 
 ### Changed